# product-service/docs/swagger.yaml
definitions:
  main.Product:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
host: localhost:3002
info:
  contact: {}
  description: API de autenticação com JWT
  title: Auth Service API
  version: "1.0"
paths:
  /api/products:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de produtos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Product'
            type: array
      security:
      - BearerAuth: []
      summary: Lista todos os produtos
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Cria um novo produto
      parameters:
      - description: Dados do Produto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria um produto
      tags:
      - products
  /api/products/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um produto pelo ID
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deleta um produto
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um produto pelo ID
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Product'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtém um produto
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um produto existente
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados Atualizados
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza um produto
      tags:
      - products
swagger: "2.0"